Add Optional Error Context  (#46)
As reported in MaxDesiatov/CoreXLSX#12 there are multiple issues where it's hard to understand why parsing of a certain XML has failed. `XMLCoder` should be able to report a snippet of a given XML of specified length around the position where the error has happened. This should be an property of type `UInt` named `errorContextLength` on `XMLDecoder` instance. An error thrown from XML parser that has line/column location should be repackaged having a context around that location of specified length.

For example, if an error was thrown indicating that there's an unexpected character at line 3, column 15 with `errorContextLength` set to `10`, a new error type should be rethrown containing 5 characters before column 15 and 5 characters after, all on line 3. Line wrapping should be handled correctly too as the context can span more than a few lines.

Also, currently there's a single use of `print` in the codebase for printing an error in `XMLStackParser.swift`. This has been removed.

* Add ErrorHandleTest
* Fix ErrorContextTest and formatting