Keyed/unkeyed boxes as structs (#36)
Unifies the representation of `T` where `T: Box` and makes it explicit where shared reference semantics is expected.

With this code performance is consistently 5-30% worse, could this be caused by copies of value types? Can this be improved when Swift 5.0 is released or whatever the version is that allows us to improve performance of value types? After reviewing related issues we'd prefer XMLCoder to be slightly slower, but with #12, #17 and #25 fixed than the opposite 😞 

* Added `SharedBox` type for on-demand reference semantics
* Remove `print` that kills performance tests
* Handle all cases with SharedBox containers
* Add RJITest back to the project file, fix the test
* Remove unused internal initializer