Add missing CFBundleVersion, bump version to 0.3.1 (#74)
Using carthage to either grab the latest prebuilt binary 0.3.0 or forcing carthage to build the latest XmlCoder results in the framework's plist missing a value for CFBundleVersion, which is required for AppStore Connect upload. Other plist keys are correctly filled in. I see it's using the $(CURRENT_PROJECT_VERSION) variable for CFBundleVersion.

> AppStore error: ERROR ITMS-90056: XMLCoder.framework is invalid. The Info.plist file is missing the required key: CFBundleVersion

Resolves #72