Attributed Intrinsic (#73)
## Overview

This PR fixes #12: decoding of unkeyed single value elements that contain attributes as reported in that issue.

## Example

```xml
<?xml version="1.0" encoding="UTF-8"?>
<foo id="123">456</foo>
``` 

```swift
private struct Foo: Codable, DynamicNodeEncoding {
    let id: String
    let value: String

    enum CodingKeys: String, CodingKey {
        case id
        case value
        // case value = "" would also work
    }

    static func nodeEncoding(forKey key: CodingKey) -> XMLEncoder.NodeEncoding {
        switch key {
        case CodingKeys.id:
            return .attribute
        default:
            return .element
        }
    }
}
```

Previously this XML example would fail to decode. This PR allows two different methods of decoding discussed in usage.

## Usage

This PR will support decoding the example XML in two cases as long as the prerequisite cases are matched.

### Prerequisites

1. No keyed child elements exist
2. Any keyed child nodes are supported Attribute types and are indicated to decode as such

### Supported cases

1. An instance var with the key `value` of any decodable type.
2. An instance var of any key that has a Decoding key of String value "value" or "".

The decoder will look for the case where an element was keyed with either "value" or "", but not both, and only one of those values (ie; no other keyed elements). It will automatically find the correct value based on the CodingKey supplied.

## Other considerations

The choice to decode as either "value" or "" keys was purely to try to support the inverse to XML version which would only work if an instance var specifically has a `CodingKey` with associated value type `String` that returns an empty string, if PR #70 is commited as-is, which adds XML coding support for unkeyed attributed value elements.

The 'value' variant was added as a simpler means to support decoding a nested unkeyed element without having to provide custom CodingKey enum for a struct. Something needed to be provided since Swift doesn't have empty string iVars `let "" : String`, isn't a valid iVar token for example, so `value` was chosen as a logical default.

## Notes

This PR is an extension of #70 , though it could be recoded to work off of `master`. The last commit in this PR is the only commit specific to this feature, though #70 provides the inverse solution of creating XML from an attributed value wrapping struct.

Coding and decoding unit tests of String and Int values are included.