Add separate lint stage to .travis.yml (#87)
The prevents linting stage from running multiple times in all parallel jobs. We're mostly going to develop with the latest version of Xcode, so we only care about linter and formatter warnings with that version. Thus, the linter stage runs first with the latest version, while actual tests run in parallel jobs of the next "test" stage.

* Travis jobs/stages experiment
* Fix pod lint warnings with Xcode 10.2
* Reorder matrix and jobs in .travis.yml
* Specify jobs/stages with YAML references
* Test .travis.yml without YAML references
* Test stages/jobs one by one
* Add install: skip to stages in .travis.yml
* Update .travis.yml
* Add codecov upload to one of the jobs
* Add back lint stage
* Rename `unit-test` stage to `test`
* Add separate Package.swift for Swift 5.0
* Add explicit top level osx_image to .travis.yml