Make value intrinsic smarter (#89)
Recent addition of value intrinsic allows XMLCoder to handle last few known edge cases, but still a lot of times the shape of an XML isn't guaranteed to be always the same. E.g. some of the attributes might be present or absent, but an element without attributes is never decoded as a `KeyedBox` even with a value intrinsic. And vice versa, a `KeyedBox` with attributes can't be decoded with a primitive type that maps to the element value. 

This was previously reported by @Inukinator as an issue in Inukinator/Read#2.

Consider this sample XML:

```xml
<container>
<foo id="123">456</foo>
<foo id="789">123</foo>
</container>
```

As of 847d754 this can't be decoded as

```swift
struct Container: Codable {
  let foo: [Int]
}
```

And this xml:

```xml
<container>
<foo>456</foo>
<foo>123</foo>
<foo>789</foo>
</container>
```

can't be decoded with

```swift
struct Container: Codable {
  struct Foo: Codable {
    let value: Int
  }

  let foo: [Foo]
}
```

This is fixed now with added optional down casts that peek into the box structure and map it correctly in these remaining edge cases.

* Refine value intrinsic tests
* Refactor XMLCoderElement.flatten, add 
* Add missing assert to 
* Fix FlattenTests error due to missing assert
* Fix new cases in AttributedIntrinsicTest