Migrate from Travis to Azure Pipelines (#111)
Unfortunately, Travis seems to be very slow and unreliable compared to Azure Pipeline builds. I've already migrated a few other OSS projects to Azure and it worked pretty great. The only thing we lose for now is binary frameworks built with Carthage attached to release assets. I don't think this is a blocker, since with the imminent release of Xcode 11 I anticipate usage of Carthage to decrease with more people migrating to SwiftPM. In addition, there's a stronger demand now for a SwiftPM binary frameworks proposal, and these scripts would have to be changed anyway to support the new XCFrameworks. Azure Pipelines does not prevent us from doing that, I'd just like to wait until all that stuff is stable enough before we start integrating this into our CI process.

* Add azure-pipelines.yml and related scripts
* Remove unused `carthage bootstrap` command
* Remove .travis.yml