Fix `trimValueWhitespaces` removing needed white-spaces (#226)
I found that the logic for trimming white-spaces (`trimValueWhitespaces`) may inadvertently remove white-spaces from inside the string value of the element, if `XMLParser` reports the content of a single element as multiple strings.  

This happens for example with non-ascii characters like German umlauts:
`Lösen` is split up and reported separately as `L` and `ösen`.
Or where we have "mixed" strings, e.g. containing a segment of Japanese or Chinese characters:
`Japanese 日本語` is split into `Japanese` and ` 日本語`.

If there's a white-space adjacent to the split, as in the Japanese example above, the white-space gets trimmed **before** it's joined together. Resulting in `Japanese日本語` without a space in the middle.

I added failing tests with the first commit.

* Add failing tests

see: https://stackoverflow.com/a/1076951/2064473

"The parser object may send the delegate several parser:foundCharacters: messages to report the characters of an element. Because string **may be only part of the total character content for the current element**, you should append it to the current accumulation of characters until the element changes."

* trim only completed string elements

Refactor `process()` function name to `trimWhitespacesIfNeeded()`

* Add additional tests for non-standard characters

for nested xml and untrimmed case

thanks to @wooj2 for the suggestion!