Update XMLCoderElement.swift
The existing implementation was wrong, when the element is empty we set an empty xml tag. We should simply not set it. 

Example :
```swift
structure Test { var Title: String, var description: String? }
```

In this case calling encode() function like this : 
```swift
    let xmlEncoder = XMLEncoder()
    xmlEncoder.keyEncodingStrategy = XMLEncoder.KeyEncodingStrategy.capitalized
    let data = try xmlEncoder.encode(self, withRootKey:"Test")
``

Result in : 
```xml
<Test><Title>"Test"</Title><Description /></Test>
```

The update I made will simply return this instead of an xml string containing an empty element : 
```xml
<Test><Title>"Test"</Title></Test>
```